Moved that bugger to the wrong method. Fixed
Really, up there is the TL;DR.
Due to onSongEnd() attempting to set it to false after calling next(), an empty Queue would set everything up to get a NullPointerException thrown. as onSongEnd() set songEnded to false, calling next() to start the next song in the Queue would cause next() to check if it was called from onSongEnd(), as it wasn't it called stop() on an alreadt stopped player.
Putting that statement way down in or after playTrack is a bad idea as well as the whole threading can sometimes cause a bit of a wait, where the right circumstances could cause onSongEnd() to be called while it still registers true on songEnded.